Changes to resources. Fixed favicon. Deleted megabyte cms because why is it here.
Functional mostly however the rich text that is parsed is not responsive (html needs to be correctly fit to container). Page title is also not responsive and at smaller screens wraps to next line. Grid is working well when the two descriptions are of similar length however does not work for longer descriptions.